Routes path and controllers structure refactore
